Use 1-pixel-wide rectangles instead of arbitrary rectangles.
This commit also uses the computed polynomial as long as it gives tight
enclosures for every pixel. An enclosure P([u,v]) of f is deemed tight if
it is almost the same as an underapproximation of [f(u),f(v)]. If not, a
new polynomial is computed on a smaller range of pixels.

If a polynomial is successful for all the pixels of the original range, a
larger range is chosen next. This makes it possible to dynamically adjust
the accuracy of the computations.